# src/main/resources/META-INF/swagger/guesto-1.0.yml
openapi: 3.0.1
info:
  title: guesto
  version: "1.0"
security:
- bearerAuth: []
tags:
- name: Events
  description: Operations related to events management
- name: Users
  description: User authentication and management
paths:
  /event:
    get:
      tags:
      - Events
      summary: List Events
      description: Retrieves a list of all events.
      operationId: listEvents
      responses:
        "200":
          description: listEvents 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      tags:
      - Events
      summary: Create Event
      description: Creates a new event with the given details.
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDTO'
        required: true
      responses:
        "200":
          description: createEvent 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
  /event/{eventId}/add-guest:
    post:
      tags:
      - Events
      summary: Add Guest to Event
      description: Adds a guest to the specified event's guest list.
      operationId: addGuestToEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestDTO'
        required: true
      responses:
        "200":
          description: addGuestToEvent 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestDTO'
  /event/{eventId}/guests:
    get:
      tags:
      - Events
      summary: List All Guests
      description: Lists all guests for the specified event.
      operationId: listAllGuests
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: listAllGuests 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GuestDTO'
  /event/{eventId}/guests/{guestId}:
    put:
      tags:
      - Events
      summary: Update Guest in Event
      description: Updates a guest in the specified event's guest list.
      operationId: updateGuestInEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: guestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestDTO'
        required: true
      responses:
        "200":
          description: updateGuestInEvent 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestDTO'
    delete:
      tags:
      - Events
      summary: Delete Guest from Event
      description: Deletes a guest from the specified event's guest list.
      operationId: deleteGuestFromEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: guestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteGuestFromEvent 200 response
          content:
            application/json:
              schema:
                type: object
  /event/{id}:
    put:
      tags:
      - Events
      summary: Update Event
      description: Updates the event with the specified ID.
      operationId: updateEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDTO'
        required: true
      responses:
        "200":
          description: updateEvent 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDTO'
    delete:
      tags:
      - Events
      summary: Delete Event
      description: Deletes the event with the specified ID.
      operationId: deleteEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteEvent 200 response
          content:
            application/json:
              schema:
                type: object
  /event/{id}/checkin:
    put:
      tags:
      - Events
      summary: Check In Guest
      description: Checks in a guest for the event with the specified ID.
      operationId: checkInGuest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int64
        required: true
      responses:
        "200":
          description: checkInGuest 200 response
          content:
            application/json:
              schema:
                type: object
  /user/list:
    get:
      tags:
      - Users
      summary: List All Users
      description: Lists all registered users.
      operationId: listUsers
      responses:
        "200":
          description: listUsers 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
  /user/login:
    post:
      tags:
      - Users
      summary: User Login
      description: Authenticates a user with the provided credentials.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
        required: true
      responses:
        "200":
          description: login 200 response
          content:
            application/json:
              schema:
                type: object
  /user/register:
    post:
      tags:
      - Users
      summary: User Registration
      description: Registers a new user with the provided details.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDTO'
        required: true
      responses:
        "200":
          description: register 200 response
          content:
            application/json:
              schema:
                type: object
  /user/{userId}:
    delete:
      tags:
      - Users
      summary: Delete User
      description: Deletes a user with the specified ID.
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deleteUser 200 response
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        eventName:
          type: string
        eventTime:
          type: string
          format: date-time
        maxGuestList:
          type: integer
          format: int32
        price:
          type: integer
        location:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        checkedInGuestsCount:
          type: integer
          format: int32
        guestList:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/GuestList'
    EventDTO:
      type: object
      properties:
        eventName:
          type: string
        eventTime:
          type: string
          format: date-time
        maxGuestList:
          type: integer
          format: int32
        price:
          type: integer
        location:
          type: string
        guestList:
          $ref: '#/components/schemas/GuestListDTO'
    Guest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        checkedIn:
          type: boolean
        additionalGuests:
          type: integer
          format: int32
        guestList:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/GuestList'
    GuestDTO:
      required:
      - additionalGuests
      - firstName
      - lastName
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        checkedIn:
          type: boolean
        additionalGuests:
          type: integer
          format: int32
        guests:
          type: array
          items:
            $ref: '#/components/schemas/GuestDTO'
    GuestList:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        event:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/Event'
        guests:
          type: array
          items:
            $ref: '#/components/schemas/Guest'
    GuestListDTO:
      required:
      - eventId
      - guests
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        guests:
          type: array
          items:
            $ref: '#/components/schemas/GuestDTO'
    LoginDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    RegisterDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT